services/ledger-api-py/requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.0
psycopg2-binary==2.9.9
pydantic==2.7.2
boto3==1.34.131
grpcio==1.64.1
grpcio-tools==1.64.1
protobuf==5.27.1
python-dotenv==1.0.1
